backend/requirements.txt
aniso8601==9.0.1
attrs==22.1.0
blinker==1.5
click==8.1.3
colorama==0.4.5
countryinfo==0.1.2
Flask==1.1.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-Excel==0.0.7
Flask-Login==0.6.2
flask-marshmallow==0.14.0
Flask-Migrate==4.0.4
flask-restx==0.5.1
Flask-SQLAlchemy==2.4.4
Flask_JWT_Extended==4.3.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==4.14.0
MarkupSafe==1.1.1
marshmallow==3.9.1
#marshmallow-sqlalchemy==0.24.1
marshmallow-sqlalchemy
openpyxl==3.0.10
pandas
pre-commit==2.20.0
#psycopg2==2.9.3
pyrsistent==0.18.1
python-dotenv
pytz==2022.2.1
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.0
Werkzeug==1.0.1
xlrd==2.0.1
